cleaned up the code. Made additional components for the word
